Move helpers up to mull impl
